% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_posteriors.R
\name{get_posteriors}
\alias{get_posteriors}
\title{Collect posteriors from VPVD stanfit object}
\usage{
get_posteriors(
  stanfit,
  group_tag = "by_group_drug$",
  subject_tag = "subject_effect$"
)
}
\arguments{
\item{stanfit}{The stanfit object you would like to extract posteriors from}

\item{group_tag}{Regex code to identify the names of the group paramaters}

\item{subject_tag}{Regex code to identify the names of the subject parameters}
}
\value{
List of lists: One list for grouped parameters and one list for individuals
}
\description{
Collect posteriors from VPVD stanfit object
}
