% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_pred_check.R
\name{post_pred_check}
\alias{post_pred_check}
\title{Posterior predictive check}
\usage{
post_pred_check(
  alpha_w = 0.002,
  alpha_l = 0.002,
  beta = 1.5,
  kappa = 0.2,
  tau = 0.1,
  upsilon = 5,
  nTrials = 2000,
  nIter = 10
)
}
\arguments{
\item{nTrials}{}
}
\description{
Posterior predictive check
}
