% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_pred_check.R
\name{prior_pred_check}
\alias{prior_pred_check}
\title{Prior Predictive Checks for the Valence Probe Visual Discrimination Task}
\usage{
prior_pred_check(
  alpha = NA,
  alpha_w = 0.002,
  alpha_l = 0.002,
  beta = 1.5,
  kappa = 0.2,
  tau = 0.1,
  upsilon = 5,
  eta = NA,
  phi = NA,
  subjID = 1,
  grouping = 1,
  nTrials = 2000
)
}
\arguments{
\item{nTrials}{}
}
\description{
Prior Predictive Checks for the Valence Probe Visual Discrimination Task
}
