% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_choices_VPVD.R
\name{simulate_choices_vpvd}
\alias{simulate_choices_vpvd}
\title{Simulates choices and probabilities in VPVD}
\usage{
simulate_choices_vpvd(
  nIter = 10,
  nTrials = 2000,
  nGroups = 1,
  alpha = NA,
  alpha_w = matrix(data = rep(0.1, nIter * nGroups), ncol = nGroups),
  alpha_l = matrix(data = rep(0.5, nIter * nGroups), ncol = nGroups),
  beta = matrix(data = rep(1.5, nIter * nGroups), ncol = nGroups),
  kappa = matrix(data = rep(0.2, nIter * nGroups), ncol = nGroups),
  tau = matrix(data = rep(0.1, nIter * nGroups), ncol = nGroups),
  upsilon = matrix(data = rep(5, nIter * nGroups), ncol = nGroups),
  eta = NA,
  phi = NA
)
}
\arguments{
\item{nTrials}{}
}
\description{
Simulates choices and probabilities in VPVD
}
