% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_correct.R
\name{summarise_correct}
\alias{summarise_correct}
\title{Summarises correct choices - faster function (no for loop)}
\usage{
summarise_correct(list)
}
\arguments{
\item{list}{}
}
\description{
Summarises correct choices - faster function (no for loop)
}
