% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_correct_by_subj.R
\name{summarise_correct_by_subj}
\alias{summarise_correct_by_subj}
\title{Summarises data by subject (not group)}
\usage{
summarise_correct_by_subj(list, groups = 1:length(unique(list$grouping)))
}
\arguments{
\item{list}{}
}
\value{
Data frame with session data for each subject and trial type
}
\description{
Summarises data by subject (not group)
}
