% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_sessions_from_list.R
\name{summarise_sessions_from_list}
\alias{summarise_sessions_from_list}
\title{Summarises the returned output list from a simulation}
\usage{
summarise_sessions_from_list(list)
}
\arguments{
\item{list}{}
}
\description{
Summarises the returned output list from a simulation
}
