% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trials}
\alias{trials}
\title{Trial sequence with 4000 trials from the VPVD task}
\format{
A data frame with 4,000 rows and 2 columns:
\describe{
\item{stim_left}{Stimulus presented to the left}
\item{stim_right}{Stimulus presented to the right}
}
}
\source{
{One of 10 trial sequences used in the original VPVD task}
}
\usage{
trials
}
\description{
An example sequence where the stimuli are coded as 0, 50 and 100.
}
\keyword{datasets}
