% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_posteriors.R
\name{visual_posteriors}
\alias{visual_posteriors}
\title{Visualise density plots from posteriors}
\usage{
visual_posteriors(params, grouping = 1:dim(params)[[1]][2])
}
\arguments{
\item{params}{These are the posterior parameters that you want to use for the simulation}

\item{grouping}{Names of the groups/conditions in the experiment.}
}
\value{
Plots a cowplot of one density plot for each parameter
}
\description{
Visualise density plots from posteriors
}
