% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_correct_by_subj.R
\name{summarise_correct_by_subj}
\alias{summarise_correct_by_subj}
\title{Summarises data by subject (not group)}
\usage{
summarise_correct_by_subj(list)
}
\arguments{
\item{list}{}
}
\description{
Summarises data by subject (not group)
}
